requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
distro==1.9.0
dotenv==0.9.9
elevenlabs==2.36.1
frozenlist==1.8.0
google-auth==2.48.0
google-genai==1.64.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
multidict==6.7.1
numpy==2.4.2
propcache==0.4.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
sounddevice==0.5.5
soundfile==0.13.1
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==15.0.1
yarl==1.22.0
pandas>=2.2.0
openpyxl>=3.1.0

torch>=2.2.0
transformers>=4.40.0
